+++ src/tests/api/institution/test_learner.py
@@ -38,8 +38,6 @@ def test_api_institution_learner(rest_api_client, user_global_admin, institution
        9) Remove a SEND Category from a learner
        10) Delete new learner
     """
-    pytest.skip("TODO: Adapt to the change from learner to user endpoints")
-
     institution_user = institution_course_test_case['user'].institutionuser
     institution_id = institution_course_test_case['institution'].id
     learner = institution_course_test_case['learner']
@@ -187,7 +185,7 @@ def test_api_institution_learner(rest_api_client, user_global_admin, institution
     # 5) Accept an informed consent for a learner
     """ ---------------------------------------------------------------------
     ACCEPT INFORMED CONSENT FOR A LEARNER:
-        POST /api/v2/institution/(int: institution_id)/user/(int: learner_id)/ic/
+        POST /api/v2/institution/(int: institution_id)/learner/(int: learner_id)/ic/
         Request JSON Object: version (string) – Informed consent version to assign
         Status Codes:
             200 OK – Ok
@@ -206,7 +204,7 @@ def test_api_institution_learner(rest_api_client, user_global_admin, institution
     # If no valid informed consent exist: create one
     if ic_body['count'] > 0:
         ic_version = ic_body['results'][0].version
-        str_path = '/api/v2/institution/{}/user/{}/ic'.format(institution_id, new_learner_id)
+        str_path = '/api/v2/institution/{}/learner/{}/ic'.format(institution_id, new_learner_id)
         str_data = {'version': ic_version}
         tests.utils.post_rest_api_client(rest_api_client, str_path, str_data,
                                          'Accept Informed Consent for a Learner',
@@ -224,7 +222,7 @@ def test_api_institution_learner(rest_api_client, user_global_admin, institution
                                          'Accept Informed Consent for a Learner',
                                          'RESPONSE: ', 200)
 
-    str_path = '/api/v2/institution/{}/user/{}/'.format(institution_id, new_learner_id)
+    str_path = '/api/v2/institution/{}/learner/{}/'.format(institution_id, new_learner_id)
     str_response = 'RESPONSE Learner ID={}:'.format(new_learner_id)
     body = tests.utils.get_rest_api_client(rest_api_client, str_path,
                                            'Read Learner Information', str_response, 200)
@@ -233,7 +231,7 @@ def test_api_institution_learner(rest_api_client, user_global_admin, institution
     # 6) Reject current Informed Consent of a learner
     """ ---------------------------------------------------------------------
     REJECT CURRENT INFORMED CONSENT OF A LEARNER:
-        DELETE /api/v2/institution/(int: institution_id)/user/(int: learner_id)/ic/
+        DELETE /api/v2/institution/(int: institution_id)/learner/(int: learner_id)/ic/
         Status Codes:
             200 OK – Ok
             400 Bad Request – Invalid information provided. The response contains the description of the errors.
@@ -242,11 +240,11 @@ def test_api_institution_learner(rest_api_client, user_global_admin, institution
     Request Headers: Authorization - JWT with Institution Admin privileges
     """
     logging.info('\n6) REJECT INFORMED CONSENT FOR A LEARNER --------------------------------------')
-    str_path = '/api/v2/institution/{}/user/{}/ic/'.format(institution_id, new_learner_id)
+    str_path = '/api/v2/institution/{}/learner/{}/ic/'.format(institution_id, new_learner_id)
     tests.utils.delete_rest_api_client(rest_api_client, str_path,
                                        'Reject Current Informed Consent of a Learner', "RESPONSE: ", 200)
 
-    str_path = '/api/v2/institution/{}/user/{}/'.format(institution_id, new_learner_id)
+    str_path = '/api/v2/institution/{}/learner/{}/'.format(institution_id, new_learner_id)
     str_response = 'RESPONSE Learner ID={}:'.format(new_learner_id)
     body = tests.utils.get_rest_api_client(rest_api_client, str_path,
                                            'Read Learner Information', str_response, 200)
